Moved domains beansprofile and roasters into src
