+++ src/beansprofile/controller.js
